Main/src-server/linux/NetIf-linux.cpp: clang build fixes, bugref:10391
svn:sync-xref-src-repo-rev: r168050

+++ src/VBox/Main/src-server/linux/NetIf-linux.cpp
@@ -1,4 +1,4 @@
-/* $Id: NetIf-linux.cpp 106061 2024-09-16 14:03:52Z knut.osmundsen@oracle.com $ */
+/* $Id: NetIf-linux.cpp 108833 2025-03-20 12:20:01Z alexander.eichner@oracle.com $ */
 /** @file
  * Main - NetIfList, Linux implementation.
  */
@@ -230,7 +230,7 @@ static int getInterfaceInfo(int iSocket, const char *pszName, PNETIFINFO pInfo)
                     pInfo->IPv6Address.au32[1] = htonl(IPv6Address.au32[1]);
                     pInfo->IPv6Address.au32[2] = htonl(IPv6Address.au32[2]);
                     pInfo->IPv6Address.au32[3] = htonl(IPv6Address.au32[3]);
-                    RTNetPrefixToMaskIPv6(uLength, &pInfo->IPv6NetMask);
+                    RTNetPrefixToMaskIPv6((int)uLength, &pInfo->IPv6NetMask);
                 }
             }
             fclose(fp);
